Дополнение к заданию 3
